Add find_provider_action 1-0-1
We need to track additional information that we did not recognize when we created 1-0-0. This includes adding the actual and max score of the selection rules, what was the name of the selection rule, and if all of the selection rule criteria were met; sort order and sort column; what was the confidence score.

Signed-off-by: Evan Duncan <evan.duncan@fieldnation.com>